Add index.html and Modify pizza_crust.html
Added the index.html to the project along with the directory leading to
the pizza_crust.html file. The pizza_crust.html was modified with HTML
boilerplate, the directory leading back to the index.html page, an image,
and all of the elements that were necessary to complete the recipe.